+++ client/pages/auth/verify-email.js
@@ -2,12 +2,12 @@ import Button from "@mui/material/Button";
 import Stack from "@mui/material/Stack";
 import Typography from "@mui/material/Typography";
 import AuthLayout from "../../components/Layouts/AuthLayout";
-import AuthHeader from "../../components/auth/AuthHeader";
+import AuthHeader from "../../components/Auth/AuthHeader";
 
 export default function VerifyEmail() {
   return (
     <AuthLayout>
-      <AuthHeader subtitle="Sign in to your account" />
+      <AuthHeader subtitle="Verify your email address" />
       <Stack spacing={4} sx={{ my: 2 }}>
         <Typography variant="h6">
           We have sent an email to ninomaeianis@hololive.com.
